Seemingly functional init
At some point with this, I no longer got a kernel panic when I started
alsamixer, but I did get routing errors in dmesg. I just removed all the
routing parsing and the properties from the device tree and went with
the default of the DAC, and then the routing errors disappears as well.
I had a working alsamixer with controls.

But then, when I rebooted the device, I had the same null pointer error
again, but this time when loading the driver, not just when starting
alsamixer...