Implement no-signal detection
It works in conjunction with ALSA SOC's anti-pop feature: the DAC's
analogue mute is disengaged when the playback device is opened, and
re-engaged when it closes.

It fixes two problems related to bad or missing clock output from the
DIR9001 when there is no source bi-phase signal:

- Undefined operation of the PCM1690, producing noise.
- Delay when raw PCM mode was running for a while without a signal
  present.

As of yet, this code is untested with actual audio output, because my
test-setup was limited, and I used an old prototype board.