Change no-signal detection to using sample rate
The previous method didn't fully work when you'd pause a PCM stream:
zeroes or 255 values would be given, and even when detecting those both
as no-signal, delay was introduced after pausing. It's a lot better to
keep PCM mode running.

So, use the measured sample rate. It's a hack though. See code comments
for why.